requirements.txt
requests
openai
dotenv
python-fasthtml
ollama
PyMuPDF
starlette